Exporter: factor out SelectObservation
